+++ slog/__init__.py
@@ -8,7 +8,7 @@ from .cli import main, print_help, print_version
 from .enums import IntEnumPlus, ReturnCodes, LogLevel
 from .logs import activate_logging, deactivate_logging, flush_logging, log_multiline
 from .paths import get_root_dir, get_tests_dir
-from .utils import consecutive, is_dunder, capture_output
+from .utils import CaptureOutputResult, consecutive, is_dunder, capture_output
 from .version import version_info
 
 #%% Constants