+++ slog/utils.py
@@ -11,11 +11,54 @@ from contextlib import contextmanager
 import doctest
 from io import StringIO
 import sys
-from typing import Any, Callable, TypeVar
+from typing import Any, Callable, Iterator, Optional, TextIO, TypeVar, Union
 import unittest
 
 #%% Constants
 _F = TypeVar("_F", bound=Callable[..., Any])
+
+
+#%% Classes
+class CaptureOutputResult:
+    r"""Class used to keep track of the standard output and error streams to assist the capture_output function."""
+
+    def __init__(
+        self,
+        stdout: Optional[Union[StringIO, TextIO]] = None,
+        stderr: Optional[Union[StringIO, TextIO]] = None,
+    ):
+        self.stdout = stdout
+        self.stderr = stderr
+
+    def close(self) -> None:
+        r"""Closes any open streams."""
+        if self.stdout:
+            self.stdout.close()
+
+        if self.stderr:
+            self.stderr.close()
+
+    def get_output(self) -> str:
+        r"""Returns what was captured in the output stream."""
+        return CaptureOutputResult.get_stream(self.stdout)
+
+    def get_error(self) -> str:
+        r"""Returns what was captured in the error stream."""
+        return CaptureOutputResult.get_stream(self.stderr)
+
+    @staticmethod
+    def get_stream(std: Optional[Union[StringIO, TextIO]]) -> str:
+        r"""Gets the contents of the given stream."""
+        if not std:
+            return ""
+
+        if isinstance(std, StringIO):
+            return std.getvalue().strip()
+
+        if isinstance(std, TextIO):
+            return "\n".join(std.readlines())
+
+        raise Exception(f"Unknown type {type(std)}")
 
 
 #%% Decorators - consecutive
@@ -90,7 +133,7 @@ def is_dunder(name: str) -> bool:
 
 #%% Functions - capture_output
 @contextmanager
-def capture_output(mode: str = "out"):
+def capture_output(mode: str = "out") -> Iterator[CaptureOutputResult]:
     r"""
     Capture the stdout and stderr streams instead of displaying to the screen.
 
@@ -98,9 +141,9 @@ def capture_output(mode: str = "out"):
     ----------
     mode : str
         Mode to use when capturing output
-            'out' captures just sys.stdout
-            'err' captures just sys.stderr
-            'all' captures both sys.stdout and sys.stderr
+            "out" captures just sys.stdout
+            "err" captures just sys.stderr
+            "all" captures both sys.stdout and sys.stderr
 
     Returns
     -------
@@ -112,14 +155,16 @@ def capture_output(mode: str = "out"):
     Notes
     -----
     #.  Written by David C. Stauffer in March 2015.
+    #.  Updated by David C. Stauffer in August 2022 based on stackoverflow answer by user rshepp.
+        See: https://stackoverflow.com/questions/73228026/python-typing-for-context-manager-and-string-literal/
 
     Examples
     --------
     >>> from slog import capture_output
-    >>> with capture_output() as out:
+    >>> with capture_output() as ctx:
     ...     print('Hello, World!')
-    >>> output = out.getvalue().strip()
-    >>> out.close()
+    >>> output = ctx.get_output()
+    >>> ctx.close()
     >>> print(output)
     Hello, World!
 
@@ -139,11 +184,11 @@ def capture_output(mode: str = "out"):
             sys.stderr = new_err
         # yield results as desired
         if mode == "out":
-            yield sys.stdout
+            yield CaptureOutputResult(stdout=sys.stdout)
         elif mode == "err":
-            yield sys.stderr
+            yield CaptureOutputResult(stderr=sys.stderr)
         elif mode == "all":
-            yield sys.stdout, sys.stderr
+            yield CaptureOutputResult(stdout=sys.stdout, stderr=sys.stderr)
     finally:
         # restore the original buffers once all results are read
         sys.stdout, sys.stderr = old_out, old_err