+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## August 2022 (by David C. Stauffer)
+
+* Reworked the capture_output function to internally use a Class to abstract the details.
+* Publish the v1.0.0 release.
+
 ## February 2022 (by David C. Stauffer)
 
 * Split into a stand-alone library using the latest standards of pyproject.toml, plus single source versioning, generic named src folder, separated tests and robust testing with pytest, mypy, black and coverage.